backend/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "omicsomics-backend"
version = "0.1.0"
description = "Backend services for the Omicsomics unified omics platform"
requires-python = ">=3.10"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "sqlalchemy[asyncio]>=2.0.30",
  "asyncpg>=0.29.0",
  "alembic>=1.13.0",
  "pydantic[email]>=2.6.0",
  "pydantic-settings>=2.2.1",
  "python-dotenv",
  "httpx",
  "aioboto3>=12.0.0",
  "boto3",
  "pandas",
  "numpy",
  "scikit-learn",
  "scanpy",
  "xgboost",
  "networkx",
  "pyyaml",
  "passlib[bcrypt]>=1.7.4",
  "python-jose[cryptography]>=3.3.0",
  "python-multipart>=0.0.9",
  "python-magic>=0.4.27",
  "aiofiles>=23.0.0",
  "aiosqlite>=0.20.0",
  "cryptography>=41.0.0",
  "celery[redis]>=5.3.0",
  "redis>=5.0.0",
  "pyteomics>=4.6.0",
  "openpyxl>=3.1.0",
  # Bioinformatics format converters (Phase 1 Week 1)
  "biopython>=1.81",
  "pybedtools>=0.9.0",
  "pyBigWig>=0.3.18",
  # Bioinformatics format converters (Phase 1 Week 2)
  "pysam>=0.21.0",
  # Bioinformatics format converters (Phase 1 Week 3)
  "scipy>=1.11.0",
  "h5py>=3.9.0",
  "rpy2>=3.5.0"
]

[project.optional-dependencies]
dev = [
  "pytest",
  "pytest-asyncio",
  "httpx",
  "pytest-cov",
  "asgi-lifespan",
  "mypy",
  "ruff",
  "pre-commit"
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--strict-markers",
    "--strict-config",
    "--showlocals",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
]

[tool.coverage.run]
source = ["app"]
omit = [
    "*/tests/*",
    "*/migrations/*",
    "*/__pycache__/*",
    "*/site-packages/*",
]

[tool.coverage.report]
precision = 2
show_missing = true
skip_covered = false
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]

[tool.ruff]
select = ["E", "F", "I", "UP", "B", "Q"]
ignore = ["E203", "E501"]
line-length = 100

[tool.mypy]
python_version = "3.11"
plugins = []
disallow_untyped_defs = true
ignore_missing_imports = true
namespace_packages = true
warn_unused_ignores = true

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]
exclude = ["tests*"]
